EID-828: renamed misspelt package name
