+++ src/main/java/uk/gov/ida/eidas/metadata/saml/SamlObjectSigner.java
@@ -1,4 +1,4 @@
-package uk.gov.ida.eidas.metatdata.saml;
+package uk.gov.ida.eidas.metadata.saml;
 
 import org.opensaml.core.xml.XMLObject;
 import org.opensaml.core.xml.io.MarshallingException;
