+++ src/test/java/uk/gov/ida/eidas/metadata/ConnectorMetadataSignerTest.java
@@ -12,7 +12,6 @@ import org.opensaml.saml.common.SignableSAMLObject;
 import org.opensaml.xmlsec.signature.Signature;
 import uk.gov.ida.common.shared.security.PrivateKeyFactory;
 import uk.gov.ida.common.shared.security.X509CertificateFactory;
-import uk.gov.ida.eidas.metatdata.ConnectorMetadataSigner;
 import uk.gov.ida.eidas.utils.FileReader;
 import uk.gov.ida.saml.core.test.TestCertificateStrings;
